+++ assess/src/codatassess/models/shared/excelstatus.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 import dataclasses
+from ..shared import excelreporttypes as shared_excelreporttypes
 from codatassess import utils
 from dataclasses_json import Undefined, dataclass_json
 from typing import Optional
@@ -13,8 +14,11 @@ from typing import Optional
 class ExcelStatus:
     r"""OK"""
     error_message: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('errorMessage'), 'exclude': lambda f: f is None }})
+    r"""Error details in case the report generation request was unsuccessful."""
     file_size: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('fileSize'), 'exclude': lambda f: f is None }})
+    r"""The file size in Bytes is populated upon successful generation of the report."""
     in_progress: Optional[bool] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('inProgress'), 'exclude': lambda f: f is None }})
+    r"""When true, the request was successful and the report is being generated. If false, the request was unsuccessful and the report is not being generated."""
     last_generated: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('lastGenerated'), 'exclude': lambda f: f is None }})
     r"""In Codat's data model, dates and times are represented using the <a class=\\"external\\" href=\\"https://en.wikipedia.org/wiki/ISO_8601\\" target=\\"_blank\\">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
 
@@ -37,8 +41,12 @@ class ExcelStatus:
     > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
     """
     last_invocation_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('lastInvocationId'), 'exclude': lambda f: f is None }})
+    r"""A unique ID generated for this request."""
     queued: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('queued'), 'exclude': lambda f: f is None }})
-    report_type: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('reportType'), 'exclude': lambda f: f is None }})
+    r"""The date and time of when a successful request was queued for the most recent report."""
+    report_type: Optional[shared_excelreporttypes.ExcelReportTypes] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('reportType'), 'exclude': lambda f: f is None }})
+    r"""The type of the report requested in the query string."""
     success: Optional[bool] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('success'), 'exclude': lambda f: f is None }})
+    r"""True if the requested report was successfully queued and false if the requested report was not able to be queued."""