+++ assess/RELEASES.md
@@ -653,3 +653,13 @@ Based on:
 - [python v0.33.1] assess
 ### Releases
 - [PyPI v0.33.1] https://pypi.org/project/codat-assess/0.33.1 - assess
+
+## 2023-08-30 00:12:26
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.74.19 (2.88.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.33.2] assess
+### Releases
+- [PyPI v0.33.2] https://pypi.org/project/codat-assess/0.33.2 - assess